Update noise_module.py
This change finds and fixes an important bug related to the time vector of the cross-correlation functions. The previous version accidentally shifts the cross-correlation functions of one sample towards the positive time axis, which is problematic for long-period studies. This bug can not be found without Dr. Xingli Fan's help!